rename package_manager_sanity_check
